% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_plot_cercle_cor_plotly.R
\name{circle_plot}
\alias{circle_plot}
\title{Correlation circle plot for PLSDA}
\usage{
circle_plot(PLSDA)
}
\arguments{
\item{PLSDA}{a PLSDA object}
}
\value{
a correlation circle
}
\description{
Correlation circle plot for PLSDA
}
\examples{
PLSDA::circle_plot(res)
}
