% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_matrix.R
\name{classification_report}
\alias{classification_report}
\title{Classification report function for PLSDA}
\usage{
classification_report(observed, predict)
}
\arguments{
\item{observed}{observed values (ytest)}

\item{predict}{predicted values (values returned with predict function)}
}
\value{
\code{Confusion_matrix} Confusion matrix
\cr
\code{report} Report with precision, recall and f1-score for each class
\cr
\code{f1_score} global f1 score 
\cr
}
\description{
Classification report function for PLSDA
}
\examples{
res=plsda_Classification_report(ytrain,ypred)
}
