% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.plot.R
\name{explanatory_variables}
\alias{explanatory_variables}
\title{Plot of explanatory variables}
\usage{
explanatory_variables(var1, var2, color)
}
\arguments{
\item{var1}{A quantitative variable}

\item{var2}{A quantitative variable}

\item{color}{Color of the points in the plot}
}
\value{
a scatter plot
}
\description{
'Plot of explanatory variables
}
\examples{
explanatory_variables(var1 = iris$Sepal.Length,var2=iris$Sepal.Width,color=iris$Species)
}
