% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_plot_indiv.R
\name{indiv_plot}
\alias{indiv_plot}
\title{Individuals plots for PLSDA}
\usage{
indiv_plot(PLSDA, axe1 = 1, axe2 = 2)
}
\arguments{
\item{PLSDA}{PLSDA object (returned with the fit function)}

\item{axe1}{1st component to observe the individuals}

\item{axe2}{2nd component to observe the individuals}
}
\value{
a plot with the 2 choosen components and the individuals on it
}
\description{
Individuals plots for PLSDA
}
