% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_split_sample.R
\name{pls.train_test_split}
\alias{pls.train_test_split}
\title{Make a train and a test sample.}
\usage{
pls.train_test_split(data, prop = 66)
}
\arguments{
\item{data}{dataset chosen by the user}

\item{prop}{Proportion of the whole dataset which is going to be in the train split. 
For exemple, if the user write prop=66, it means that an amount of 66% of the dataset will be for the train data. 
prop=66 is the parameter by default}
}
\value{
Train and test data
}
\description{
'This function makes a train and test samples based on a proportion chosen by the user
}
\examples{
train_test=pls.train_test_split(iris)
train_test=pls.train_test_split(iris,prop=70)
}
