% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_cv.r
\name{plsda.cross_validation}
\alias{plsda.cross_validation}
\title{Make a cross-validation on on dataset to select the best sample}
\usage{
plsda.cross_validation(formula, data, ncomp, nfolds = 5)
}
\arguments{
\item{formula}{formula an object of class formula" (or one that can be coerced to that class): a symbolic description of the model
to be fitted.}

\item{data}{dataframe containing the variables in the model.}

\item{ncomp}{Number of components extracted in NIPALS algorithm.}

\item{nfold}{Number of folds in the cross-validation (5 by default)}
}
\value{
\code{bestmodel} the best model with the cross-validation
\cr
\code{bestfscore} the fscore of the best model
\cr
}
\description{
Make a cross-validation on on dataset to select the best sample
}
\examples{

cv=plsda.cross_validation(Species~.,data=iris,2)
pred=plsda.predict(Species~.,data=iris,2,nfolds=10)
}
