% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_nipals.R
\name{plsda.nipals}
\alias{plsda.nipals}
\title{NIPALS Algorithm for Partial Least Square Discriminant Analysis}
\usage{
plsda.nipals(X, Y, data, ncomp)
}
\arguments{
\item{data}{is the dataframe containing the the variables in the model.}

\item{ncomp}{is the number of components for X.}

\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model
to be fitted. See specification of the formula in the 'Details' section.}
}
\value{
An object of class 'PLSDA' is a list containing at least the following components :

\code{X} the original dataframe of the predictors
\cr
\code{Y} the original variable to predict
\cr
\code{Yloadings} the matrix of loadings for Y.
\cr
\code{Yscores} the matrix of components for Y.
\cr
\code{Xloadings} the matrix of loadings for X.
\cr
\code{Xloading.weights} the matrix of weights of the loadings of X.
\cr
\code{Xscores}the matrix of components of X.
\cr
\code{TrainPlsData}the PLS-DA training data set.
\cr
\code{R2} the coefficient of determination of the PLS-DA.
}
\description{
This function performs NIPALS algorithm for PLS-DA regression.
}
\examples{
pls.t1<-plsda.pls(Species~.,data = iris, ncomp = 2)
pls.t1<-plsda.pls(Species~.,data = iris, ncomp = 2, center = TRUE)
}
