% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_predict.R
\name{plsda.predict}
\alias{plsda.predict}
\title{Predict y for a given x with a PLSDA model}
\usage{
plsda.predict(PLSDA, newdata, type = "class")
}
\arguments{
\item{PLSDA}{PLSDA type model}

\item{newdata}{Dataset for which we want to make predictions}

\item{type}{class (by default) if we want to return the class variable 
posterior if we want to know the probabily of belonging of each class for the new dataset}
}
\value{
\code{Ypred} if type = "posterior". Probabily of belonging of each class for the new dataset
\cr
\code{pred} if type = "class". Return the class variable 
\cr
}
\description{
Predict y for a given x with a PLSDA model
}
\examples{

pred=plsda.predict(mod,newdata)
pred=plsda.predict(mod,newdata,type="posterior")

}
