% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_print_summary.R
\name{plsda.print}
\alias{plsda.print}
\title{print function for PLSDA}
\usage{
plsda.print(PLS)
}
\arguments{
\item{PLSDA}{PLSDA object (returned with fit function)}
}
\value{
a print function to observe the coefficients and intercept
obtained with the fit function
}
\description{
print function for PLSDA
}
\examples{
plsda.print(res)
}
