% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_vip.R
\name{plsda.vip}
\alias{plsda.vip}
\title{Title}
\usage{
plsda.vip(PLS, threshold = 0.8)
}
\arguments{
\item{threshold}{criterion of variable selection 0.8 by default}

\item{A}{PLSDA type model}
}
\value{
\code{newX} New dataset with variable selection
\cr
\code{vip} Selected variable in the model
\cr
}
\description{
Title
}
\examples{
res=plsda.vip=(objPLDS)
res=plsda.vip=(objPLDS,threshold=1)
}
