% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_print.R
\name{print.PLSDA}
\alias{print.PLSDA}
\title{print function for PLSDA}
\usage{
\method{print}{PLSDA}(PLS)
}
\arguments{
\item{PLSDA}{PLSDA object (returned with fit function)}
}
\value{
a print function to observe the coefficients and intercept
obtained with the fit function
}
\description{
print function for PLSDA
}
\examples{
PLSDA::print(res)
}
