% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{PLSDA Scree plot}
\usage{
scree_plot(acp)
}
\arguments{
\item{acp}{A ACP model}
}
\value{
a scree plot which shows which component select
}
\description{
Make a scree plot for the component selection
}
\examples{
plsda_scree_plot(res)
}
