% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.plsda}
\alias{summary.plsda}
\title{Summary function for PLSDA}
\usage{
\method{summary}{plsda}(PLSDA, Xtest, ytest)
}
\arguments{
\item{PLSDA}{PLSDA object (obtained with fit function)}

\item{Xtest}{dataframe with the sames variables of PLSDA$X to predict the values}

\item{ytest}{classes that correspond to the values of Xtest, to compare with the predicted values}
}
\value{
a print function to observe 
- the coefficients and intercept obtained with the fit function
- a classification report including Confusion matrix,
  Report with precision, recall and f1-score for each class,
  global f1 score
}
\description{
Summary function for PLSDA
}
\examples{
PLSDA::summary(res)
}
