% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_vip.R
\name{vip}
\alias{vip}
\title{Variable Importance in the Projection}
\usage{
vip(PLSDA, threshold = 0.8)
}
\arguments{
\item{PLSDA}{A PLSDA type model}

\item{threshold}{criterion of variable selection 0.8 by default}
}
\value{
\code{newX} New dataset with variable selection
\cr
\code{vip} Importance of each variable for each composants
\cr
}
\description{
Variable Importance in the Projection
}
\examples{
res=PLSDA::vip=(PLSDA)
res=PLSDA::vip=(PLSDA,threshold=1)
}
