% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_plot_cercle_cor_plotly.R
\name{circle_plot}
\alias{circle_plot}
\title{Correlation circle plot for PLSDA}
\usage{
circle_plot(acp)
}
\arguments{
\item{acp}{a ACP type variable}
}
\value{
a correlation circle
}
\description{
Correlation circle plot for PLSDA
}
\examples{
circle.plot(res)
}
