% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_plot_cercle_cor_plotly.R
\name{circle_plot}
\alias{circle_plot}
\title{Correlation circle plot for PLSDA}
\usage{
circle_plot(PLSDA, axe1 = 1, axe2 = 2)
}
\arguments{
\item{PLSDA}{a PLSDA object}

\item{axe1}{1st component to observe the individuals}

\item{axe2}{2nd component to observe the individuals}
}
\value{
a correlation circle
}
\description{
Correlation circle plot for PLSDA
}
\examples{
PLSDA::circle_plot(res)
}
