% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda_scree_plot.R
\name{scree_plot}
\alias{scree_plot}
\title{PLSDA Scree plot}
\usage{
scree_plot(PLSDA)
}
\arguments{
\item{PLSDA}{A PLSDA object}
}
\value{
a scree plot which shows which component select
}
\description{
Make a scree plot for the component selection
}
\examples{
PLSDA::scree_plot(res)
}
